Quick fix for overflow on 32bit systems

+++ eth/p2p/rlpx_protocols/les/private/les_types.nim
@@ -85,7 +85,7 @@ type
     reqCount*, maxReqCount*: int
     sumWeigth*: int
 
-    rechargingRate*: int
+    rechargingRate*: int64
     totalRechargedUnits*: int
     totalRechargingPower*: int
